+++ tests/deprecated_api/test_remove_1-8.py
@@ -32,10 +32,10 @@ from pytorch_lightning.plugins.training_type.single_device import SingleDevicePl
 from pytorch_lightning.plugins.training_type.single_tpu import SingleTPUPlugin
 from pytorch_lightning.plugins.training_type.tpu_spawn import TPUSpawnPlugin
 from pytorch_lightning.trainer.states import RunningStage
-from pytorch_lightning.utilities import rank_zero_warn
 from pytorch_lightning.utilities.apply_func import move_data_to_device
 from pytorch_lightning.utilities.enums import DeviceType, DistributedType
 from pytorch_lightning.utilities.imports import _TORCHTEXT_LEGACY
+from pytorch_lightning.utilities.rank_zero import rank_zero_warn
 from tests.helpers.boring_model import BoringDataModule, BoringModel
 from tests.helpers.runif import RunIf
 from tests.helpers.torchtext_utils import get_dummy_torchtext_data_iterator
@@ -413,3 +413,37 @@ def test_v1_8_0_on_configure_sharded_model(tmpdir):
         match="The `on_configure_sharded_model` callback hook was deprecated in v1.6 and will be removed in v1.8."
     ):
         trainer.fit(model)
+
+
+def test_v1_8_0_rank_zero_imports():
+
+    import warnings
+
+    from pytorch_lightning.utilities.distributed import rank_zero_debug, rank_zero_info
+    from pytorch_lightning.utilities.warnings import LightningDeprecationWarning, rank_zero_deprecation, rank_zero_warn
+
+    with pytest.deprecated_call(
+        match="pytorch_lightning.utilities.distributed.rank_zero_debug has been deprecated in v1.6"
+        " and will be removed in v1.8."
+    ):
+        rank_zero_debug("foo")
+    with pytest.deprecated_call(
+        match="pytorch_lightning.utilities.distributed.rank_zero_info has been deprecated in v1.6"
+        " and will be removed in v1.8."
+    ):
+        rank_zero_info("foo")
+    with pytest.deprecated_call(
+        match="pytorch_lightning.utilities.warnings.rank_zero_warn has been deprecated in v1.6"
+        " and will be removed in v1.8."
+    ):
+        rank_zero_warn("foo")
+    with pytest.deprecated_call(
+        match="pytorch_lightning.utilities.warnings.rank_zero_deprecation has been deprecated in v1.6"
+        " and will be removed in v1.8."
+    ):
+        rank_zero_deprecation("foo")
+    with pytest.deprecated_call(
+        match="pytorch_lightning.utilities.warnings.LightningDeprecationWarning has been deprecated in v1.6"
+        " and will be removed in v1.8."
+    ):
+        warnings.warn("foo", LightningDeprecationWarning, stacklevel=5)